+++ src/app.js
@@ -1,6 +1,6 @@
 import Fastify from 'fastify';
-import { getDB, saveDB } from './utils/db/index.js';
-import { v4 } from 'uuid';
+import { general } from './services/general/index.js';
+import { createTodo } from './services/todos/create-todos.js';
 
 const prefix = '/api';
 
@@ -8,47 +8,10 @@ export async function build () {
   // initialize fastify
   const fastify = Fastify({ logger: true });
 
-  fastify.get(prefix, async (request, reply) => {
-    return { success: true };
-  });
+  fastify.get(prefix, general);
 
   // create todo
-  fastify.post(`${prefix}/todo`, async (request, reply) => {
-    const { body } = request;
-    const { title, description, isDone = false } = body;
-    const db = await getDB();
-
-    const id = v4();
-
-    const todo = {
-      title,
-      description,
-      isDone,
-      createdDate: new Date().getTime(),
-      updateDate: new Date().getTime()
-    };
-
-    db.todos[id] = todo;
-
-    await saveDB(db);
-
-    /**
-     * const newObj = {
-     *  id
-     * }
-     *
-     * for(const key in todo){
-     * newObj[key] = todo[key]
-     * }
-     *
-     * return newObj;
-     */
-
-    return {
-      id,
-      ...todo // elements of todo
-    };
-  });
+  fastify.post(`${prefix}/todo`, createTodo);
 
   return fastify;
 }
