+++ src/services/todos/create-todos.js
@@ -1,0 +1,38 @@
+import { getDB, saveDB } from '../../utils/db/index.js';
+import { v4 } from 'uuid';
+
+export const createTodo = async (request, reply) => {
+  const { body } = request;
+  const { title, description, isDone = false } = body;
+  const db = await getDB();
+
+  const id = v4();
+
+  const todo = {
+    title,
+    description,
+    isDone,
+    createdDate: new Date().getTime(),
+    updateDate: new Date().getTime()
+  };
+
+  db.todos[id] = todo;
+
+  await saveDB(db);
+
+  /**
+     * const newObj = {
+     *  id
+     * }
+     *
+     * for(const key in todo){
+     * newObj[key] = todo[key]
+     * }
+     *
+     * return newObj;
+     */
+  return {
+    id,
+    ...todo // elements of todo
+  };
+};
